NG bhork first version

+++ uarchsim/vht.h
@@ -9,5 +9,5 @@ class vht{
     bool get_value(uint64_t pc, uint64_t bhr, uint64_t* value);
     bool increment_os_branch_count(uint64_t pc, uint64_t bhr);
     bool decrement_os_branch_count(uint64_t pc, uint64_t bhr);
-    bool update_branch_difference(uint64_t pc, uint64_t bhr, int64_t diff);
+    bool update_branch_difference(uint64_t pc, uint64_t bhr, int32_t diff);
 };
